eco-index-main/requirements.txt
albumentations
datasets
evaluate
huggingface-hub
matplotlib
numpy
Pillow
playwright
scikit-learn
torch
torchvision
tqdm
transformers
